FormatHtmlOperation: Make it modular
