+++ src/main/kotlin/logic/document/SettingsBasedDocumentProcessor.kt
@@ -3,10 +3,9 @@ package logic.document
 import kmtt.models.entry.Entry
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.flow.launchIn
-import kotlinx.coroutines.flow.onEach
 import logic.abstracts.AbstractProgress
 import logic.document.operations.*
+import logic.document.operations.html.FormatHtmlOperation
 import logic.document.operations.media.SaveMediaOperation
 import logic.document.operations.media.modules.IDownloadModule
 import logic.document.operations.media.modules.ImageDownloadModule
