+++ src/main/kotlin/logic/document/operations/html/modules/IHtmlFormatModule.kt
@@ -1,0 +1,7 @@
+package logic.document.operations.html.modules
+
+import org.jsoup.nodes.Document
+
+interface IHtmlFormatModule {
+    fun process(document: Document): Document
+}
